refactor(dqm_ingest): moved functions to read and write ref data to ref module
